Chaplin-brunch: Update to NPM, Fix bugs
