+++ examples/chaplin-brunch/.gitignore
@@ -1,8 +1,1 @@
-bower_components/
-!bower_components/todomvc-common/
-
-# NPM packages folder.
 node_modules/
-
-# Brunch folder for temporary files.
-tmp/
