+++ examples/chaplin-brunch/app/assets/index.html
@@ -3,8 +3,14 @@
 	<head>
 		<meta charset="utf-8">
 		<title>Chaplin &amp; Brunch • TodoMVC</title>
-		<link rel="stylesheet" href="../bower_components/todomvc-common/base.css">
-		<link rel="stylesheet" href="../bower_components/todomvc-common/base.js">
+		<link rel="stylesheet" href="../node_modules/todomvc-common/base.css">
+		<link rel="stylesheet" href="../node_modules/todomvc-app-css/index.css">
+		<script src="../node_modules/todomvc-common/base.js"></script>
+		<script src="../node_modules/exoskeleton/exoskeleton.js"></script>
+		<script src="../node_modules/backbone.nativeview/backbone.nativeview.js"></script>
+		<script src="../node_modules/backbone.localstorage/backbone.localStorage.js"></script>
+		<script src="vendor.js"></script>
+		<script src="../node_modules/chaplin/chaplin.js"></script>
 		<script src="app.js"></script>
 		<script>require('initialize');</script>
 	</head>
