+++ examples/chaplin-brunch/app/views/templates/footer.hbs
@@ -14,7 +14,4 @@
     <a href="#/completed">Completed</a>
   </li>
 </ul>
-<button id="clear-completed">
-  Clear completed
-  <span id="completed-count"></span>
-</button>
+<button id="clear-completed">Clear completed <span id="completed-count"></span></button>
