--- examples/chaplin-brunch/bower.json
@@ -1,20 +1,0 @@
-{
-  "name": "todomvc-chaplin-brunch",
-  "version": "0.0.0",
-  "dependencies": {
-    "todomvc-common": "~0.3.0",
-    "chaplin": "~1.0.0",
-    "exoskeleton": "~0.6.3",
-    "backbone.localStorage": "git://github.com/paulmillr/Backbone.localStorage.git#9c225b37bdea4ac21d4b2445fa8962fe74e3175b"
-  },
-  "overrides": {
-    "todomvc-common": {
-      "main": "bg.png"
-    },
-    "chaplin": {
-      "dependencies": {
-        "exoskeleton": "*"
-      }
-    }
-  }
-}
