+++ examples/chaplin-brunch/config.coffee
@@ -2,7 +2,9 @@ exports.config =
   # See http://brunch.readthedocs.org/en/latest/config.html for documentation.
   files:
     javascripts:
-      joinTo: 'app.js'
+      joinTo:
+        'app.js': /^app/
+        'vendor.js': /^vendor/
 
     stylesheets:
       joinTo: 'app.css'
